+++ src/MiscellaneousCopulas/WCopula.jl
@@ -31,4 +31,4 @@ StatsBase.corkendall(::WCopula) = [1 -1; -1 1]
 StatsBase.corspearman(::WCopula) = [1 -1; -1 1]
 
 # Subsetting colocated
-SubsetCopula(::WCopula, ::NTuple{p, Int64}) where {p} = (p==2 ? C : error("WCopula only defined for p=2"))
+SubsetCopula(::WCopula, ::NTuple{p, Int}) where {p} = (p==2 ? C : error("WCopula only defined for p=2"))
